+++ clients/admin-ui/src/features/datastore-connections/ConnectionFilters.tsx
@@ -3,52 +3,29 @@ import React from "react";
 import { useDispatch, useSelector } from "react-redux";
 
 import { SearchLineIcon } from "../common/Icon";
-import SystemTypeMenu from "./ConnectionDropdown";
-import ConnectionStatusMenu from "./ConnectionStatusMenu";
-import { DisabledStatus, SystemType, TestingStatus } from "./constants";
 import {
   selectDatastoreConnectionFilters,
-  setDisabledStatus,
   setSearch,
-  setSystemType,
-  setTestingStatus,
 } from "./datastore-connection.slice";
+import ConnectionTypeFilter from "./filters/ConnectionTypeFilter";
+import DisabledStatusFilter from "./filters/DisabledStatusFilter";
+import SystemTypeFilter from "./filters/SystemTypeFilter";
+import TestingStatusFilter from "./filters/TestingStatusFilter";
 
 const useConstantFilters = () => {
   const filters = useSelector(selectDatastoreConnectionFilters);
   const dispatch = useDispatch();
   const handleSearchChange = (event: React.ChangeEvent<HTMLInputElement>) =>
     dispatch(setSearch(event.target.value));
-  const handleSystemTypeChange = (value: string) =>
-    dispatch(setSystemType(value));
-  const handleTestingStatusChange = (value: string) =>
-    dispatch(setTestingStatus(value));
-  const handleDisabledStatusChange = (value: string) =>
-    dispatch(setDisabledStatus(value));
 
   return {
     handleSearchChange,
-    handleSystemTypeChange,
-    handleTestingStatusChange,
-    handleDisabledStatusChange,
     ...filters,
   };
 };
 
 const ConnectionFilters: React.FC = () => {
-  const {
-    handleSearchChange,
-    handleSystemTypeChange,
-    handleTestingStatusChange,
-    handleDisabledStatusChange,
-    search,
-    // eslint-disable-next-line @typescript-eslint/naming-convention
-    system_type,
-    // eslint-disable-next-line @typescript-eslint/naming-convention
-    test_status,
-    // eslint-disable-next-line @typescript-eslint/naming-convention
-    disabled_status,
-  } = useConstantFilters();
+  const { handleSearchChange, search } = useConstantFilters();
   return (
     <Stack direction="row" spacing={4} mb={6}>
       <InputGroup size="sm">
@@ -66,26 +43,10 @@ const ConnectionFilters: React.FC = () => {
           onChange={handleSearchChange}
         />
       </InputGroup>
-      <ConnectionStatusMenu />
-      <SystemTypeMenu
-        title="System Type"
-        filterOptions={Object.values(SystemType)}
-        value={system_type}
-        setValue={handleSystemTypeChange}
-      />
-      <SystemTypeMenu
-        title="Testing Status"
-        filterOptions={Object.values(TestingStatus)}
-        value={test_status}
-        setValue={handleTestingStatusChange}
-      />
-
-      <SystemTypeMenu
-        title="Status"
-        filterOptions={Object.values(DisabledStatus)}
-        value={disabled_status}
-        setValue={handleDisabledStatusChange}
-      />
+      <ConnectionTypeFilter />
+      <SystemTypeFilter />
+      <TestingStatusFilter />
+      <DisabledStatusFilter />
     </Stack>
   );
 };