+++ CHANGELOG.md
@@ -22,6 +22,9 @@ The types of changes are:
 * Erasure support for Salesforce [#888](https://github.com/ethyca/fidesops/pull/888)
 * Access and erasure support for Sendgrid contacts endpoint [#883](https://github.com/ethyca/fidesops/pull/883)
 
+### Changed
+* Users should be able to click on the full field of a dropdown-type filter to open up the dropdown [#747](https://github.com/ethyca/fidesops/pull/903)
+
 ### Breaking Changes
 
 * Update fidesops to use bcrypt for hashing [#876](https://github.com/ethyca/fidesops/pull/876)